GCP_personal_project/bts-review-app-source-code/requirements.txt
asn1crypto==0.24.0
configparser==3.5.0b2
cryptography==2.6.1
entrypoints==0.3
enum34==1.1.6
ipaddress==1.0.17
keyring==17.1.1
keyrings.alt==3.1.1
pycrypto==2.6.1

pyxdg==0.25
SecretStorage==2.3.1
six==1.12.0
uWSGI==2.0.20

Flask>=2.0.1
google-cloud>=0.34.0
google-cloud-core>=1.7.1
google-cloud-pubsub>=2.6.1
google-cloud-datastore>=2.1.4
google-cloud-spanner>=3.6.0
google-cloud-storage>=1.41.0
google-cloud-language>=2.2.0
grpcio>=1.38.1
grpcio-gcp>=0.2.2
grpcio-tools>=1.38.1
